requirements.txt
certifi==2025.8.3
influxdb-client==1.36.1
OpenOPC-DA==1.5.1
Pyro5==5.15
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pywin32==225
reactivex==4.0.4
serpent==1.41
six==1.17.0
typing_extensions==4.5.0
urllib3==1.26.16
